last Update

+++ src/test/java/it/academy/andrey/ponomarev/garfield/project/driver/DriverSetup.java
@@ -1,4 +1,4 @@
-package garfield.driver;
+package it.academy.andrey.ponomarev.garfield.project.driver;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
@@ -7,7 +7,7 @@ import org.openqa.selenium.chrome.ChromeOptions;
 
 import java.time.Duration;
 
-import static garfield.constants.Constants.IMPLICITLY;
+import static it.academy.andrey.ponomarev.garfield.project.constants.Constants.IMPLICITLY;
 
 public class DriverSetup {
     public static WebDriver driver;
@@ -15,10 +15,13 @@ public class DriverSetup {
     public static WebDriver createDriver() {
         ChromeOptions options = new ChromeOptions();
         options.addArguments("--remote-allow-origins=*");
-        driver = new ChromeDriver(options);
-        driver.manage().window().maximize();
-        driver.manage().deleteAllCookies();
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(IMPLICITLY));
+        if (driver==null) {
+            WebDriverManager.chromedriver().setup();
+            driver = new ChromeDriver(options);
+            driver.manage().window().maximize();
+            driver.manage().deleteAllCookies();
+            driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(IMPLICITLY));
+        }
         return driver;
     }
